+++ src/pools/worker.ts
@@ -172,6 +172,13 @@ export interface WorkerUsage {
 }
 
 /**
+ * Worker strategy data.
+ */
+export interface StrategyData {
+  virtualTaskEndTimestamp?: number
+}
+
+/**
  * Worker interface.
  */
 export interface IWorker {
@@ -227,6 +234,11 @@ export interface IWorkerNode<Worker extends IWorker, Data = unknown> {
    * Worker usage statistics.
    */
   readonly usage: WorkerUsage
+  /**
+   * Worker strategy data.
+   * This is used to store data that is specific to the worker choice strategy.
+   */
+  strategyData?: StrategyData
   /**
    * Message channel (worker_threads only).
    */